+++ sqldelight-compiler/src/test/kotlin/app/cash/sqldelight/core/async/AsyncQueriesTypeTest.kt
@@ -52,6 +52,8 @@ class AsyncQueriesTypeTest {
     Truth.assertThat(result.compilerOutput).containsKey(database)
     Truth.assertThat(result.compilerOutput[database].toString()).isEqualTo(
       """
+      |@file:Suppress("RedundantVisibilityModifier", "ASSIGNED_VALUE_IS_NEVER_READ")
+      |
       |package com.example.testmodule
       |
       |import app.cash.sqldelight.SuspendingTransacterImpl
@@ -120,6 +122,8 @@ class AsyncQueriesTypeTest {
     Truth.assertThat(result.compilerOutput).containsKey(dataQueries)
     Truth.assertThat(result.compilerOutput[dataQueries].toString()).isEqualTo(
       """
+      |@file:Suppress("RedundantVisibilityModifier", "ASSIGNED_VALUE_IS_NEVER_READ")
+      |
       |package com.example
       |
       |import app.cash.sqldelight.Query