+++ sqldelight-compiler/src/test/migration-interface-fixtures/varying-query-migration-packages/output/com/example/queries/DataQueries.kt
@@ -1,3 +1,5 @@
+@file:Suppress("RedundantVisibilityModifier", "ASSIGNED_VALUE_IS_NEVER_READ")
+
 package com.example.queries
 
 import app.cash.sqldelight.ExecutableQuery